added nglPath::SetExtension better doc for nglSyncEvent fixed software renderer on win32
git-svn-id: svn://libnui.net/nui/trunk/nui3@3750 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ include/nuiVertex.h
@@ -45,7 +45,7 @@ inline uint32 NUI_RGBA(uint8 R, uint8 G, uint8 B, uint8 A)
 #ifdef _WIN32_
   return  ( (A << 24) + (R << 16) + (G << 8) + B );
 #else
-  return  ( (((uint32)A) << 24) + (((uint32)R) << 16) + (((uint32)G) << 8) + ((uint32)B) );
+  return  ( (A << 24) + (R << 16) + (G << 8) + B );
 #endif
 }
 #endif
